moved message struct to separate file, newRouter to router.go
